rpm: skip qubes-mgmt-salt-dom0-update dependency in VM
Currently there is no way to run a Salt state system-wide from a VM,
so applying updates from qui-updates widget wont work. We need a
solution for that, but until then, exclude the dependency.

Keep the widget included anyway, as the update notification is still a
useful thing to have.

QubesOS/qubes-issues#4186