Do not require qubes-mgmt-salt-dom0-update directly
It is going to be installed anyway, and we don't need specific version
dependency anymore, as it's included in the base Qubes version already.

But drop the dependency to to allow installing the package in VM too.
builderv2 does not support setting different macros for dom0 and vm
builds (at least not yet), so the dependency cannot be made conditional.

This mostly reverts 78bf15c70900aac28a7ea981da466d01112180e4

Temporary workaround due to QubesOS/qubes-issues#7950